APF-874: Changed the appfuse-maven-plugin's groupId to org.appfuse.
In order to get the appfuse-maven-plugin uploaded to Maven's Central repository, it can't have org.codehaus.mojo as its groupId. The original reason I started using this groupId is because I couldn't get the "appfuse" prefix to work w/o it. Now the prefix seems to work fine (at least locally). 

http://jira.codehaus.org/browse/MAVENUPLOAD-1701 

NOTE: We realize it makes good sense to change the groupIds to be more specific than "org.appfuse" for many of AppFuse's modules. However, it seems somewhat risky to do this late in the 2.0 game. If we do decide to change them, we'll likely do it for 2.1. Please comment on the following thread if you have any opinions on this: 

http://www.nabble.com/Thinking-out-loud-about-groupIds-tf4387544s2369.html

git-svn-id: https://appfuse.dev.java.net/svn/appfuse/trunk@2926 2aeb74e6-0f1c-0410-991c-92d95bb9df37