APF-874: Changed groupId from org.appfuse back to org.codehaus.mojo.
If we want to allow users to type "appfuse:*", they'll need to register the plugin in their ~/.m2/settings.xml with this change. Therefore, I'm reverting to org.codehaus.mojo as its easier for users. We won't be able to upload AMP to Maven's central repo until we move to Codehaus, but that's OK IMO. 

After more IRC discussions, it appears it *is* possible to distribute this plugin w/o requiring users to modify settings.xml, but the fact that we're already using org.codehaus.mojo causes that one to always be picked up first.

Since we're approaching 2.0, I'm going to leave it as org.codehaus.org for now. We can rename it to something else when Maven allows the search order to be configurable or we're ready to remove it completely from AppFuse's repository. Another option is to get everything in Maven's central repository and disable AppFuse's repository so the new one is found instead.

git-svn-id: https://appfuse.dev.java.net/svn/appfuse/trunk@2933 2aeb74e6-0f1c-0410-991c-92d95bb9df37